+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [3.1.6] - 2026-06-05
+
+- feat: added `Address#toJSON`
+
 ### [3.1.5] - 2026-05-26
 
 - fix: `Group.format()` now emits the RFC 5322 trailing `;`
@@ -75,3 +79,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 [3.1.3]: https://github.com/haraka/email-address/releases/tag/v3.1.3
 [3.1.4]: https://github.com/haraka/email-address/releases/tag/v3.1.4
 [3.1.5]: https://github.com/haraka/email-address/releases/tag/v3.1.5
+[3.1.6]: https://github.com/haraka/email-address/releases/tag/v3.1.6